fix : cannot use radius

+++ internals/shape/circle.go
@@ -5,16 +5,16 @@ import (
 	"math"
 )
 
-func circleArea(r float32) float32 {
+func circleArea(r float64) float64 {
 	return math.Pi * r * r
 }
 
-func circumference(r float32) float32 {
-	const diameter = float32(2) * math.Pi
+func circumference(r float64) float64 {
+	const diameter = float64(2) * math.Pi
 	return diameter * r
 }
 
-func PrintCircleResult(r float32) {
+func PrintCircleResult(r float64) {
 	fmt.Print("Keliling Lingkaran : ")
 	fmt.Println(circumference(r))
 	fmt.Print("Luas Lingkaran : ")
